Data.Swagger

Description

Swagger™ is a project used to describe and document RESTful APIs.

The Swagger specification defines a set of files required to describe such an API. These files can then be used by the Swagger-UI project to display the API and Swagger-Codegen to generate clients in various languages. Additional utilities can also take advantage of the resulting files, such as testing tools.

How to use this library

This section explains how to use this library to work with Swagger specification.

Monoid instances

Virtually all types representing Swagger specification have Monoid instances. The Monoid type class provides two methods — mempty and mappend.

In this library you can use mempty for a default/empty value. For instance:

>>> encode (mempty :: Swagger)
"{\"swagger\":\"2.0\",\"info\":{\"version\":\"\",\"title\":\"\"}}"

As you can see some spec properties (e.g. "version") are there even when the spec is empty. That is because these properties are actually required ones.

You should always override the default (empty) value for these properties, although it is not strictly necessary:

>>> encode mempty { _infoTitle = "Todo API", _infoVersion = "1.0" }
"{\"version\":\"1.0\",\"title\":\"Todo API\"}"

You can merge two values using mappend or its infix version (<>):

>>> encode $ mempty { _infoTitle = "Todo API" } <> mempty { _infoVersion = "1.0" }
"{\"version\":\"1.0\",\"title\":\"Todo API\"}"

This can be useful for combining specifications of endpoints into a whole API specification:

-- /account subAPI specification
accountAPI :: Swagger

-- /task subAPI specification
taskAPI :: Swagger

-- while API specification is just a combination
-- of subAPIs' specifications
api :: Swagger
api = accountAPI <> taskAPI

Lenses and prisms

Since Swagger has a fairly complex structure, lenses and prisms are used to work comfortably with it. In combination with Monoid instances, lenses make it fairly simple to construct/modify any part of the specification:

>>> :{
encode $ (mempty :: Swagger)
  & definitions .~ [ ("User", mempty & type_ .~ SwaggerString) ]
  & paths .~
    [ ("/user", mempty & get ?~ (mempty
        & produces ?~ MimeList ["application/json"]
        & at 200 ?~ ("OK" & _Inline.schema ?~ Ref (Reference "User"))
        & at 404 ?~ "User info not found")) ]
:}
"{\"swagger\":\"2.0\",\"info\":{\"version\":\"\",\"title\":\"\"},\"definitions\":{\"User\":{\"type\":\"string\"}},\"paths\":{\"/user\":{\"get\":{\"responses\":{\"404\":{\"description\":\"User info not found\"},\"200\":{\"schema\":{\"$ref\":\"#/definitions/User\"},\"description\":\"OK\"}},\"produces\":[\"application/json\"]}}}}"

In the snippet above we declare an API with a single path /user. This path provides method GET which produces application/json output. It should respond with code 200 and body specified by schema User which is defined in definitions property of swagger specification. Alternatively it may respond with code 404 meaning that user info is not found.

For convenience, swagger2 uses classy field lenses. It means that field accessor names can be overloaded for different types. One such common field is description. Many components of a Swagger specification can have descriptions, and you can use the same name for them:

>>> encode $ (mempty :: Response) & description .~ "No content"
"{\"description\":\"No content\"}"
>>> :{
encode $ (mempty :: Schema)
  & type_       .~ SwaggerBoolean
  & description ?~ "To be or not to be"
:}
"{\"type\":\"boolean\",\"description\":\"To be or not to be\"}"

ParamSchema is basically the base schema specification and many types contain it (see HasParamSchema). So for convenience, all ParamSchema fields are transitively made fields of the type that has it. For example, you can use type_ to access SwaggerType of Header schema without having to use paramSchema:

>>> encode $ (mempty :: Header) & type_ .~ SwaggerNumber
"{\"type\":\"number\"}"

Additionally, to simplify working with Response, both Operation and Responses have direct access to it via at code. Example:

>>> :{
encode $ (mempty :: Operation)
  & at 404 ?~ "Not found"
:}
"{\"responses\":{\"404\":{\"description\":\"Not found\"}}}"

You might've noticed that type_ has an extra underscore in its name compared to, say, description field accessor. This is because type is a keyword in Haskell. A few other field accessors are modified in this way:

in_, type_, default_ (as keywords);

maximum_ and minimum_ (as conflicting with Prelude);

enum_ (as conflicting with Control.Lens).

Schema specification

ParamSchema and Schema are the two core types for data model specification.

ParamSchema t specifies all the common properties, available for every data schema. The t parameter imposes some restrictions on type and items properties (see SwaggerType and SwaggerItems).

Schema is used for request and response bodies and allows specifying objects with properties in addition to what ParamSchema provides.

In most cases you will have a Haskell data type for which you would like to define a corresponding schema. To facilitate this use case swagger2 provides two classes for schema encoding. Both these classes provide means to encode types as Swagger schemas.

ToParamSchema is intended to be used for primitive API endpoint parameters, such as query parameters, headers and URL path pieces. Its corresponding value-encoding class is ToHttpApiData (from http-api-data package).

ToSchema is used for request and response bodies and mostly differ from primitive parameters by allowing objects/mappings in addition to primitive types and arrays. Its corresponding value-encoding class is ToJSON (from aeson package).

While lenses and prisms make it easy to define schemas, it might be that you don't need to: ToSchema and ToParamSchema classes both have default Generic-based implementations!

ToSchema default implementation is also aligned with ToJSON default implementation with the only difference being for sum encoding. ToJSON defaults sum encoding to defaultTaggedObject, while ToSchema defaults to something which corresponds to ObjectWithSingleField. This is due to defaultTaggedObject behavior being hard to specify in Swagger.

Here's an example showing ToJSON–ToSchema correspondance:

>>> data Person = Person { name :: String, age :: Integer } deriving Generic
>>> instance ToJSON Person
>>> instance ToSchema Person
>>> encode (Person "David" 28)
"{\"age\":28,\"name\":\"David\"}"
>>> encode $ toSchema (Proxy :: Proxy Person)
"{\"required\":[\"name\",\"age\"],\"type\":\"object\",\"properties\":{\"age\":{\"type\":\"integer\"},\"name\":{\"type\":\"string\"}}}"

Re-exports

module Data.Swagger.Lens

module Data.Swagger.ParamSchema

module Data.Swagger.Schema

Swagger specification

data Swagger

This is the root document object for the API specification.

Constructors

Swagger

Fields

_swaggerInfo :: Info

Provides metadata about the API. The metadata can be used by the clients if needed.

_swaggerHost :: Maybe Host

The host (name or ip) serving the API. It MAY include a port. If the host is not included, the host serving the documentation is to be used (including the port).

_swaggerBasePath :: Maybe FilePath

The base path on which the API is served, which is relative to the host. If it is not included, the API is served directly under the host. The value MUST start with a leading slash (/).

_swaggerSchemes :: Maybe [Scheme]

The transfer protocol of the API. If the schemes is not included, the default scheme to be used is the one used to access the Swagger definition itself.

_swaggerConsumes :: MimeList

A list of MIME types the APIs can consume. This is global to all APIs but can be overridden on specific API calls.

_swaggerProduces :: MimeList

A list of MIME types the APIs can produce. This is global to all APIs but can be overridden on specific API calls.

_swaggerPaths :: HashMap FilePath PathItem

The available paths and operations for the API. Holds the relative paths to the individual endpoints. The path is appended to the basePath in order to construct the full URL.

_swaggerDefinitions :: Definitions Schema

An object to hold data types produced and consumed by operations.

_swaggerParameters :: Definitions Param

An object to hold parameters that can be used across operations. This property does not define global parameters for all operations.

_swaggerResponses :: Definitions Response

An object to hold responses that can be used across operations. This property does not define global responses for all operations.

_swaggerSecurityDefinitions :: Definitions SecurityScheme

Security scheme definitions that can be used across the specification.

_swaggerSecurity :: [SecurityRequirement]

A declaration of which security schemes are applied for the API as a whole. The list of values describes alternative security schemes that can be used (that is, there is a logical OR between the security requirements). Individual operations can override this definition.

_swaggerTags :: Set Tag

A list of tags used by the specification with additional metadata. The order of the tags can be used to reflect on their order by the parsing tools. Not all tags that are used by the Operation Object must be declared. The tags that are not declared may be organized randomly or based on the tools' logic. Each tag name in the list MUST be unique.

_swaggerExternalDocs :: Maybe ExternalDocs

Additional external documentation.

data Host

The host (name or ip) serving the API. It MAY include a port.

Constructors

Host

Fields

_hostName :: HostName

Host name.

_hostPort :: Maybe PortNumber

Optional port.

data Scheme

The transfer protocol of the API.

Constructors

Http

Https

Ws

Wss

Info types

data Info

The object provides metadata about the API. The metadata can be used by the clients if needed, and can be presented in the Swagger-UI for convenience.

Constructors

Info

Fields

_infoTitle :: Text

The title of the application.

_infoDescription :: Maybe Text

A short description of the application. GFM syntax can be used for rich text representation.

_infoTermsOfService :: Maybe Text

The Terms of Service for the API.

_infoContact :: Maybe Contact

The contact information for the exposed API.

_infoLicense :: Maybe License

The license information for the exposed API.

_infoVersion :: Text

Provides the version of the application API (not to be confused with the specification version).

data Contact

Contact information for the exposed API.

Constructors

Contact

Fields

_contactName :: Maybe Text

The identifying name of the contact person/organization.

_contactUrl :: Maybe URL

The URL pointing to the contact information.

_contactEmail :: Maybe Text

The email address of the contact person/organization.

data License

License information for the exposed API.

Constructors

License

Fields

_licenseName :: Text

The license name used for the API.

_licenseUrl :: Maybe URL

A URL to the license used for the API.

PathItem

data PathItem

Describes the operations available on a single path. A PathItem may be empty, due to ACL constraints. The path itself is still exposed to the documentation viewer but they will not know which operations and parameters are available.

Constructors

PathItem

Fields

_pathItemGet :: Maybe Operation

A definition of a GET operation on this path.

_pathItemPut :: Maybe Operation

A definition of a PUT operation on this path.

_pathItemPost :: Maybe Operation

A definition of a POST operation on this path.

_pathItemDelete :: Maybe Operation

A definition of a DELETE operation on this path.

_pathItemOptions :: Maybe Operation

A definition of a OPTIONS operation on this path.

_pathItemHead :: Maybe Operation

A definition of a HEAD operation on this path.

_pathItemPatch :: Maybe Operation

A definition of a PATCH operation on this path.

_pathItemParameters :: [Referenced Param]

A list of parameters that are applicable for all the operations described under this path. These parameters can be overridden at the operation level, but cannot be removed there. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location.

Operations

data Operation

Describes a single API operation on a path.

Constructors

Operation

Fields

_operationTags :: Set TagName

A list of tags for API documentation control. Tags can be used for logical grouping of operations by resources or any other qualifier.

_operationSummary :: Maybe Text

A short summary of what the operation does. For maximum readability in the swagger-ui, this field SHOULD be less than 120 characters.

_operationDescription :: Maybe Text

A verbose explanation of the operation behavior. GFM syntax can be used for rich text representation.

_operationExternalDocs :: Maybe ExternalDocs

Additional external documentation for this operation.

_operationOperationId :: Maybe Text

Unique string used to identify the operation. The id MUST be unique among all operations described in the API. Tools and libraries MAY use the it to uniquely identify an operation, therefore, it is recommended to follow common programming naming conventions.

_operationConsumes :: Maybe MimeList

A list of MIME types the operation can consume. This overrides the consumes. Just [] MAY be used to clear the global definition.

_operationProduces :: Maybe MimeList

A list of MIME types the operation can produce. This overrides the produces. Just [] MAY be used to clear the global definition.

_operationParameters :: [Referenced Param]

A list of parameters that are applicable for this operation. If a parameter is already defined at the PathItem, the new definition will override it, but can never remove it. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location.

_operationResponses :: Responses

The list of possible responses as they are returned from executing this operation.

_operationSchemes :: Maybe [Scheme]

The transfer protocol for the operation. The value overrides schemes.

_operationDeprecated :: Maybe Bool

Declares this operation to be deprecated. Usage of the declared operation should be refrained. Default value is False.

_operationSecurity :: [SecurityRequirement]

A declaration of which security schemes are applied for this operation. The list of values describes alternative security schemes that can be used (that is, there is a logical OR between the security requirements). This definition overrides any declared top-level security. To remove a top-level security declaration, Just [] can be used.

data Tag

Allows adding meta data to a single tag that is used by Operation. It is not mandatory to have a Tag per tag used there.

Constructors

Tag

Fields

_tagName :: TagName

The name of the tag.

_tagDescription :: Maybe Text

A short description for the tag. GFM syntax can be used for rich text representation.

_tagExternalDocs :: Maybe ExternalDocs

Additional external documentation for this tag.

type TagName = Text

Tag name.

Types and formats

data SwaggerType t where

Constructors

SwaggerString :: SwaggerType t

SwaggerNumber :: SwaggerType t

SwaggerInteger :: SwaggerType t

SwaggerBoolean :: SwaggerType t

SwaggerArray :: SwaggerType t

SwaggerFile :: SwaggerType ParamOtherSchema

SwaggerNull :: SwaggerType Schema

SwaggerObject :: SwaggerType Schema

type Format = Text

type Definitions = HashMap Text

A list of definitions that can be used in references.

data CollectionFormat t where

Determines the format of the array.

Constructors

CollectionCSV :: CollectionFormat t

CollectionSSV :: CollectionFormat t

CollectionTSV :: CollectionFormat t

CollectionPipes :: CollectionFormat t

CollectionMulti :: CollectionFormat ParamOtherSchema

Parameters

data Param

Describes a single operation parameter. A unique parameter is defined by a combination of a name and location.

Constructors

Param

Fields

_paramName :: Text

The name of the parameter. Parameter names are case sensitive.

_paramDescription :: Maybe Text

A brief description of the parameter. This could contain examples of use. GFM syntax can be used for rich text representation.

_paramRequired :: Maybe Bool

Determines whether this parameter is mandatory. If the parameter is in "path", this property is required and its value MUST be true. Otherwise, the property MAY be included and its default value is False.

_paramSchema :: ParamAnySchema

Parameter schema.

data ParamAnySchema

Constructors

ParamBody (Referenced Schema)

ParamOther ParamOtherSchema

data ParamOtherSchema

Constructors

ParamOtherSchema

Fields

_paramOtherSchemaIn :: ParamLocation

The location of the parameter.

_paramOtherSchemaAllowEmptyValue :: Maybe Bool

Sets the ability to pass empty-valued parameters. This is valid only for either ParamQuery or ParamFormData and allows you to send a parameter with a name only or an empty value. Default value is False.

_paramOtherSchemaParamSchema :: ParamSchema ParamOtherSchema

data ParamLocation

Constructors

ParamQuery
Parameters that are appended to the URL. For example, in /items?id=###, the query parameter is id.

ParamHeader
Custom headers that are expected as part of the request.

ParamPath
Used together with Path Templating, where the parameter value is actually part of the operation's URL. This does not include the host or base path of the API. For example, in items{itemId}, the path parameter is itemId.

ParamFormData
Used to describe the payload of an HTTP request when either application/x-www-form-urlencoded or multipart/form-data are used as the content type of the request (in Swagger's definition, the consumes property of an operation). This is the only parameter type that can be used to send files, thus supporting the ParamFile type. Since form parameters are sent in the payload, they cannot be declared together with a body parameter for the same operation. Form parameters have a different format based on the content-type used (for further details, consult http://www.w3.org/TR/html401/interact/forms.html#h-17.13.4).

type ParamName = Text

data Header

Constructors

Header

Fields

_headerDescription :: Maybe Text

A short description of the header.

_headerParamSchema :: ParamSchema Header

type HeaderName = Text

data Example

Constructors

Example

Fields

getExample :: Map MediaType Value

Schemas

data ParamSchema t

Constructors

ParamSchema

Fields

_paramSchemaDefault :: Maybe Value

Declares the value of the parameter that the server will use if none is provided, for example a "count" to control the number of results per page might default to 100 if not supplied by the client in the request. (Note: "default" has no meaning for required parameters.) Unlike JSON Schema this value MUST conform to the defined type for this parameter.

_paramSchemaType :: SwaggerType t

_paramSchemaFormat :: Maybe Format

_paramSchemaItems :: Maybe (SwaggerItems t)

_paramSchemaMaximum :: Maybe Scientific

_paramSchemaExclusiveMaximum :: Maybe Bool

_paramSchemaMinimum :: Maybe Scientific

_paramSchemaExclusiveMinimum :: Maybe Bool

_paramSchemaMaxLength :: Maybe Integer

_paramSchemaMinLength :: Maybe Integer

_paramSchemaPattern :: Maybe Pattern

_paramSchemaMaxItems :: Maybe Integer

_paramSchemaMinItems :: Maybe Integer

_paramSchemaUniqueItems :: Maybe Bool

_paramSchemaEnum :: Maybe [Value]

_paramSchemaMultipleOf :: Maybe Scientific

data Schema

Constructors

Schema

Fields

_schemaTitle :: Maybe Text

_schemaDescription :: Maybe Text

_schemaRequired :: [ParamName]

_schemaAllOf :: Maybe [Schema]

_schemaProperties :: HashMap Text (Referenced Schema)

_schemaAdditionalProperties :: Maybe Schema

_schemaDiscriminator :: Maybe Text

_schemaReadOnly :: Maybe Bool

_schemaXml :: Maybe Xml

_schemaExternalDocs :: Maybe ExternalDocs

_schemaExample :: Maybe Value

_schemaMaxProperties :: Maybe Integer

_schemaMinProperties :: Maybe Integer

_schemaParamSchema :: ParamSchema Schema

data NamedSchema

A Schema with an optional name. This name can be used in references.

Constructors

NamedSchema

Fields

_namedSchemaName :: Maybe Text

_namedSchemaSchema :: Schema

data SwaggerItems t where

Items for SwaggerArray schemas.

SwaggerItemsPrimitive should be used only for query params, headers and path pieces. The CollectionFormat t parameter specifies how elements of an array should be displayed. Note that fmt in SwaggerItemsPrimitive fmt schema specifies format for elements of type schema. This is different from the original Swagger's Items Object.

SwaggerItemsObject should be used to specify homogenous array Schemas.

SwaggerItemsArray should be used to specify tuple Schemas.

Constructors

SwaggerItemsPrimitive :: Maybe (CollectionFormat t) -> ParamSchema t -> SwaggerItems t

SwaggerItemsObject :: Referenced Schema -> SwaggerItems Schema

SwaggerItemsArray :: [Referenced Schema] -> SwaggerItems Schema

data Xml

Constructors

Xml

Fields

_xmlName :: Maybe Text

Replaces the name of the element/attribute used for the described schema property. When defined within the SwaggerItems (items), it will affect the name of the individual XML elements within the list. When defined alongside type being array (outside the items), it will affect the wrapping element and only if wrapped is true. If wrapped is false, it will be ignored.

_xmlNamespace :: Maybe Text

The URL of the namespace definition. Value SHOULD be in the form of a URL.

_xmlPrefix :: Maybe Text

The prefix to be used for the name.

_xmlAttribute :: Maybe Bool

Declares whether the property definition translates to an attribute instead of an element. Default value is False.

_xmlWrapped :: Maybe Bool

MAY be used only for an array definition. Signifies whether the array is wrapped (for example, <books><book><book></books>) or unwrapped (<book><book>). Default value is False. The definition takes effect only when defined alongside type being array (outside the items).

Responses

data Responses

A container for the expected responses of an operation. The container maps a HTTP response code to the expected response. It is not expected from the documentation to necessarily cover all possible HTTP response codes, since they may not be known in advance. However, it is expected from the documentation to cover a successful operation response and any known errors.

Constructors

Responses

Fields

_responsesDefault :: Maybe (Referenced Response)

The documentation of responses other than the ones declared for specific HTTP response codes. It can be used to cover undeclared responses.

_responsesResponses :: HashMap HttpStatusCode (Referenced Response)

Any HTTP status code can be used as the property name (one property per HTTP status code). Describes the expected response for those HTTP status codes.

data Response

Describes a single response from an API Operation.

Constructors

Response

Fields

_responseDescription :: Text

A short description of the response. GFM syntax can be used for rich text representation.

_responseSchema :: Maybe (Referenced Schema)

A definition of the response structure. It can be a primitive, an array or an object. If this field does not exist, it means no content is returned as part of the response. As an extension to the Schema Object, its root type value may also be "file". This SHOULD be accompanied by a relevant produces mime-type.

_responseHeaders :: HashMap HeaderName Header

A list of headers that are sent with the response.

_responseExamples :: Maybe Example

An example of the response message.

type HttpStatusCode = Int

Security

data SecurityScheme

Constructors

SecurityScheme

Fields

_securitySchemeType :: SecuritySchemeType

The type of the security scheme.

_securitySchemeDescription :: Maybe Text

A short description for security scheme.

data SecuritySchemeType

Constructors

SecuritySchemeBasic

SecuritySchemeApiKey ApiKeyParams

SecuritySchemeOAuth2 OAuth2Params

newtype SecurityRequirement

Lists the required security schemes to execute this operation. The object can have multiple security schemes declared in it which are all required (that is, there is a logical AND between the schemes).

Constructors

SecurityRequirement

Fields

getSecurityRequirement :: HashMap Text [Text]

API key

data ApiKeyParams

Constructors

ApiKeyParams

Fields

_apiKeyName :: Text

The name of the header or query parameter to be used.

_apiKeyIn :: ApiKeyLocation

The location of the API key.

data ApiKeyLocation

The location of the API key.

Constructors

ApiKeyQuery

ApiKeyHeader

OAuth2

data OAuth2Params

Constructors

OAuth2Params

Fields

_oauth2Flow :: OAuth2Flow

The flow used by the OAuth2 security scheme.

_oauth2Scopes :: HashMap Text Text

The available scopes for the OAuth2 security scheme.

data OAuth2Flow

Constructors

OAuth2Implicit AuthorizationURL

OAuth2Password TokenURL

OAuth2Application TokenURL

OAuth2AccessCode AuthorizationURL TokenURL

type AuthorizationURL = Text

The authorization URL to be used for OAuth2 flow. This SHOULD be in the form of a URL.

type TokenURL = Text

The token URL to be used for OAuth2 flow. This SHOULD be in the form of a URL.

External documentation

data ExternalDocs

Allows referencing an external resource for extended documentation.

Constructors

ExternalDocs

Fields

_externalDocsDescription :: Maybe Text

A short description of the target documentation. GFM syntax can be used for rich text representation.

_externalDocsUrl :: URL

The URL for the target documentation.

References

newtype Reference

A simple object to allow referencing other definitions in the specification. It can be used to reference parameters and responses that are defined at the top level for reuse.

Constructors

Reference

Fields

getReference :: Text

data Referenced a

Constructors

Ref Reference

Inline a

Miscellaneous

newtype MimeList

Constructors

MimeList

Fields

getMimeList :: [MediaType]

newtype URL

Constructors

URL

Fields

getUrl :: Text