Data.Swagger.Internal.ParamSchema

Documentation

binaryParamSchema :: ParamSchema t

Default schema for binary data (any sequence of octets).

byteParamSchema :: ParamSchema t

Default schema for binary data (base64 encoded).

passwordParamSchema :: ParamSchema t

Default schema for password string. "password" format is used to hint UIs the input needs to be obscured.

class ToParamSchema a where

Convert a type into a plain ParamSchema.

An example type and instance:

{-# LANGUAGE OverloadedStrings #-}   -- allows to write Text literals

import Control.Lens

data Direction = Up | Down

instance ToParamSchema Direction where
  toParamSchema = mempty
     & type_ .~ SwaggerString
     & enum_ .~ [ "Up", "Down" ]

Instead of manually writing your ToParamSchema instance you can use a default generic implementation of toParamSchema.

To do that, simply add deriving Generic clause to your datatype and declare a ToParamSchema instance for your datatype without giving definition for toParamSchema.

For instance, the previous example can be simplified into this:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics (Generic)

data Direction = Up | Down deriving Generic

instance ToParamSchema Direction

Minimal complete definition

Nothing

Methods

toParamSchema :: proxy a -> ParamSchema t

Convert a type into a plain parameter schema.

>>> encode $ toParamSchema (Proxy :: Proxy Integer)
"{\"type\":\"integer\"}"

toParamSchemaBoundedIntegral :: forall proxy a t. (Bounded a, Integral a) => proxy a -> ParamSchema t

Default plain schema for Bounded, Integral types.

>>> encode $ toParamSchemaBoundedIntegral (Proxy :: Proxy Int8)
"{\"maximum\":127,\"minimum\":-128,\"type\":\"integer\"}"

timeParamSchema :: String -> ParamSchema t

genericToParamSchema :: forall proxy a t. (Generic a, GToParamSchema (Rep a)) => SchemaOptions -> proxy a -> ParamSchema t

A configurable generic ParamSchema creator.

>>> :set -XDeriveGeneric
>>> data Color = Red | Blue deriving Generic
>>> encode $ genericToParamSchema defaultSchemaOptions (Proxy :: Proxy Color)
"{\"type\":\"string\",\"enum\":[\"Red\",\"Blue\"]}"

class GToParamSchema f where

Methods

gtoParamSchema :: SchemaOptions -> proxy f -> ParamSchema t -> ParamSchema t

class GEnumParamSchema f where

Methods

genumParamSchema :: SchemaOptions -> proxy f -> ParamSchema t -> ParamSchema t

data Proxy3 a b c

Constructors

Proxy3