Data.Swagger.Internal.Schema.Validation

Description

Validate JSON values with Swagger Schema.

Documentation

validateToJSON :: forall a. (ToJSON a, ToSchema a) => a -> [ValidationError]

Validate ToJSON instance matches ToSchema for a given value. This can be used with QuickCheck to ensure those instances are coherent:

validateToJSON (x :: Int) == []

NOTE: validateToJSON does not perform string pattern validation. See validateToJSONWithPatternChecker.

validateToJSONWithPatternChecker :: forall a. (ToJSON a, ToSchema a) => (Pattern -> Text -> Bool) -> a -> [ValidationError]

Validate ToJSON instance matches ToSchema for a given value and pattern checker. This can be used with QuickCheck to ensure those instances are coherent.

For validation without patterns see validateToJSON.

type ValidationError = String

Validation error message.

data Result a

Validation result type.

Constructors

Failed [ValidationError]
Validation failed with a list of error messages.

Passed a
Validation passed.

data Config

Validation configuration.

Constructors

Config

Fields

configPatternChecker :: Pattern -> Text -> Bool

Pattern checker for _paramSchemaPattern validation.

configDefinitions :: Definitions Schema

Schema definitions in scope to resolve references.

defaultConfig :: Config

Default Config:

defaultConfig = Config
  { configPatternChecker = \_pattern _str -> True
  , configDefinitions    = mempty
  }

newtype Validation s a

Value validation.

Constructors

Validation

Fields

runValidation :: Config -> s -> Result a

withConfig :: (Config -> Validation s a) -> Validation s a

withSchema :: (s -> Validation s a) -> Validation s a

invalid :: String -> Validation schema a

Issue an error message.

valid :: Validation schema ()

Validation passed.

check :: Lens' s (Maybe a) -> (a -> Validation s ()) -> Validation s ()

Validate schema's property given a lens into that property and property checker.

sub :: t -> Validation t a -> Validation s a

Validate same value with different schema.

sub_ :: Getting a s a -> Validation a r -> Validation s r

Validate same value with a part of the original schema.

withRef :: Reference -> (Schema -> Validation s a) -> Validation s a

Validate value against a schema given schema reference and validation function.

validateWithSchemaRef :: Referenced Schema -> Value -> Validation s ()

validateWithSchema :: Value -> Validation Schema ()

Validate JSON Value with Swagger Schema.

validateWithParamSchema :: Value -> Validation (ParamSchema t) ()

Validate JSON Value with Swagger ParamSchema.

validateInteger :: Scientific -> Validation (ParamSchema t) ()

validateNumber :: Scientific -> Validation (ParamSchema t) ()

validateString :: Text -> Validation (ParamSchema t) ()

validateArray :: Vector Value -> Validation (ParamSchema t) ()

validateObject :: HashMap Text Value -> Validation Schema ()

validateEnum :: Value -> Validation (ParamSchema t) ()

validateSchemaType :: Value -> Validation Schema ()

validateParamSchemaType :: Value -> Validation (ParamSchema t) ()